ERROR AND SUCCESS MSG ON LOGIN,
NOT FOUND PAGE,
ADJUSTED STYLING FOR COMPONENTS TO FIT ON PAGE,
