+++ Clientapp/src/baseURL.js
@@ -1,5 +1,4 @@
-//const baseURL = "https://localhost:44300";
-const baseURL = "http://68b0e1ae6611.ngrok.io";
+const baseURL = "https://localhost:44300";
+//const baseURL = "http://68b0e1ae6611.ngrok.io";
 
-
-export default baseURL;  //comment to test commit
+export default baseURL; //comment to test commit
