+++ Clientapp/src/components/ReusableComp/ActivityLog/getInitials.js
@@ -1,0 +1,6 @@
+export default (name = '') => name
+  .replace(/\s+/, ' ')
+  .split(' ')
+  .slice(0, 2)
+  .map((v) => v && v[0].toUpperCase())
+  .join('');
