--- Clientapp/src/components/bodyAdmin/getTickets/index.js
@@ -1,69 +1,0 @@
-import React, { useState } from 'react';
-import {
-  Box,
-  Container,
-  Grid,
-  makeStyles
-} from '@material-ui/core';
-//import { Pagination } from '@material-ui/lab';
-import Page from '../Page';
-//import Toolbar from './Toolbar';
-import ProductCard from './ProductCard';
-import data from './data';
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    backgroundColor: theme.palette.background.dark,
-    minHeight: '100%',
-    paddingBottom: theme.spacing(3),
-    paddingTop: theme.spacing(3)
-  },
-  productCard: {
-    height: '100%'
-  }
-}));
-
-const ProductList = () => {
-  const classes = useStyles();
-  const [products] = useState(data);
-
-  return (
-    <Page
-      className={classes.root}
-      title="Products"
-    >
-      <Container maxWidth={false}>
-        {/* <Toolbar /> */}
-        <Box mt={3}>
-          <Grid
-            container
-            spacing={3}
-          >
-            {products.map((product) => (
-              <Grid
-                item
-                key={product.id}
-                lg={4}
-                md={6}
-                xs={12}
-              >
-                <ProductCard
-                  className={classes.productCard}
-                  product={product}
-                />
-              </Grid>
-            ))}
-          </Grid>
-        </Box>
-        <Box
-          mt={3}
-          display="flex"
-          justifyContent="center"
-        >
-        </Box>
-      </Container>
-    </Page>
-  );
-};
-
-export default ProductList;
