+++ Clientapp/src/module/actions.js
@@ -1,15 +1,27 @@
 import baseURL from "../../src/baseURL";
+import { loginSuccessMsg } from "./actions";
 
 export const STORE_CREDS = "STORE_CREDS";
 export const SIGN_OUT = "SIGN_OUT";
 export const LOGIN_FAIL = "LOGIN_FAIL";
+export const ERROR_CLEAR = "ERROR_CLEAR";
+export const LOGIN_SUCCESS_MSG = "LOGIN_SUCCESS_MSG";
 
-export function loginAction(username, token, isAdmin, isLoggedIn) {
+export function loginAction(
+  username,
+  token,
+  isAdmin,
+  firstName,
+  lastName,
+  isLoggedIn
+) {
   return {
     type: STORE_CREDS,
     username,
     token,
     isAdmin,
+    firstName,
+    lastName,
     isLoggedIn,
   };
 }
@@ -27,6 +39,19 @@ export function logoutAction() {
   };
 }
 
+export function loginMsg(show) {
+  return {
+    type: LOGIN_SUCCESS_MSG,
+    show,
+  };
+}
+
+export function setErrorToNull() {
+  return {
+    type: ERROR_CLEAR,
+  };
+}
+
 export function loginThunk(username, password) {
   return function (dispatch) {
     return fetch(`${baseURL}/api/auth/login`, {
@@ -38,15 +63,27 @@ export function loginThunk(username, password) {
         if (response.ok) {
           return response.json();
         } else {
-          dispatch(loginFail("Error status::: ", response.status));
+          //dispatch(loginFail("Error status::: ", response.status));
           console.log("Response not okay", response.status);
         }
       })
       .then((json) => {
-        dispatch(loginAction(username, json.token, json.isAdmin, true));
+        dispatch(
+          loginAction(
+            username,
+            json.token,
+            json.isAdmin,
+            json.firstName,
+            json.lastName,
+            true
+          )
+        );
+
         sessionStorage.setItem("userToken", JSON.stringify(json.token));
+        dispatch(loginMsg(true));
       })
       .catch((error) => {
+        dispatch(loginFail("Login Error."));
         console.error("error from login ::", error);
       });
   };
