+++ SLU.CS.LAB/Controllers/AuthController.cs
@@ -81,7 +81,9 @@ namespace SLU.CS.LAB.Controllers
             return Ok(new
             {
                 token = tokenHandler.WriteToken(token),
-                userFromRepo.IsAdmin
+                userFromRepo.IsAdmin,
+                userFromRepo.FirstName,
+                userFromRepo.LastName
             });
         }
     }
